Fixing tests for workflows
